backend/requirements.txt
tensorflow==2.15.0
tflite-support==0.4.4
fastapi==0.104.1
uvicorn==0.24.0
gdown==5.1.0
matplotlib==3.8.2
scikit-learn==1.3.2
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
